:hammer: Refactor: Filter, DownArror 리팩토링
