+++ src/components/Community/ArticleSection/ArticleSection.js
@@ -1,11 +1,14 @@
 import './ArticleSection.scss';
 import Filter from 'components/common/Filter/Filter';
+import PrioritySelector from 'components/common/Filter/Selector/PrioritySelector';
 import ArticleList from './ArticleList/ArticleList';
 
 function ArticleSection() {
   return (
     <div className="article-section">
-      <Filter />
+      <Filter>
+        <PrioritySelector />
+      </Filter>
       <ArticleList />
     </div>
   );
