+++ src/components/Diary/Modal/ChallengeList/WeekSelector.js
@@ -3,7 +3,7 @@ import './WeekSelector.scss';
 import { useState, useEffect } from 'react';
 import * as utils from 'lib/utils/diary';
 
-import DownArrow from '../DownArrow';
+import DownArrow from 'components/common/SelectArrow/DownArrow';
 
 function WeekSelector() {
   const date = new Date();
