+++ src/components/common/Filter/Selector/Button/DropDown/Options.js
@@ -1,13 +1,14 @@
 import './Options.scss';
 import Option from './Option';
 
-function Options({ handleParameter, list, close }) {
+function Options({ handleParameter, list, close, parameterName }) {
   return (
     <div className="filter-options">
       <ul>
         {list?.map(elem => (
           <li key={elem}>
             <Option
+              parameterName={parameterName}
               value={elem}
               handleParameter={handleParameter}
               close={close}
